bin/: directory to store some external commands making them internal, to use only in myshell.
adduser.sh: Added a feature to display adding a new user form
in the middle of the screen using 'tput' command.

login.sh: Added a feature to display logging in form in the middle
of the screen using 'tput' command.

myshell.sh: Finished with the basic requirements, with
several extra features such as history command and
accessing certain commands from history with '!<command#>'.
'ls' to display the files and directories.
Probably more features will be added.